move navigator endpoints to REPORTS #223, update naming #231
