+++ tests/src/views/test_file_view.py
@@ -257,124 +257,3 @@ def test_list_posts_filter(client, stixify_file, more_files, filters, expected_i
     assert {post["id"] for post in resp.data["files"]} == set(expected_ids)
     assert resp.data["total_results_count"] == len(expected_ids)
     api_schema['/api/v1/files/']['GET'].validate_response(Transport.get_st_response(resp))
-
-
-
-
-@pytest.mark.django_db
-def test_list_attack_navigator__has_data(client, stixify_file, api_schema, navigator_data):
-    post = models.File.objects.get(pk="dcbeb240-8dd6-4892-8e9e-7b6bda30e454")
-    post.txt2stix_data = {
-        "navigator_layer": navigator_data
-    }
-    post.save()
-
-    resp = client.get(
-        "/api/v1/files/dcbeb240-8dd6-4892-8e9e-7b6bda30e454/attack-navigator/",
-        data=None,
-        content_type="application/json",
-    )
-    assert resp.status_code == 200, resp.content
-    assert resp.data == {"ics": True, "mobile": True, "enterprise": False}
-    api_schema["/api/v1/files/{file_id}/attack-navigator/"][
-        "GET"
-    ].validate_response(Transport.get_st_response(resp))
-
-
-@pytest.mark.django_db
-def test_retrieve_attack_navigator__no_data(client, stixify_file, api_schema):
-    post = models.File.objects.get(pk="dcbeb240-8dd6-4892-8e9e-7b6bda30e454")
-    post.txt2stix_data = {}
-    post.save()
-
-    resp = client.get(
-        "/api/v1/files/dcbeb240-8dd6-4892-8e9e-7b6bda30e454/attack-navigator/ics/",
-        data=None,
-        content_type="application/json",
-    )
-    assert resp.status_code == 404, resp.content
-    json.loads(resp.content) == {
-        "code": 404,
-        "details": {"error": "no navigator for this domain", "domains": []},
-        "message": "Not Found",
-    }
-    api_schema["/api/v1/files/{file_id}/attack-navigator/{attack_domain}/"][
-        "GET"
-    ].validate_response(Transport.get_st_response(resp))
-
-
-@pytest.fixture
-def navigator_data():
-    return [
-        {
-            "version": "4.5",
-            "name": "Basic ICS Layer",
-            "domain": "ics-attack",
-            "description": "A simple example ICS layer using only required fields.",
-            "techniques": [{"techniqueID": "T0887"}, {"techniqueID": "T0853"}],
-        },
-        {
-            "version": "4.5",
-            "name": "Minimal Mobile Layer",
-            "domain": "mobile-attack",
-            "description": "A basic mobile layer with minimal configuration.",
-            "techniques": [{"techniqueID": "T1406"}, {"techniqueID": "T1475"}],
-        },
-    ]
-
-
-@pytest.mark.django_db
-def test_retrieve_attack_navigator__has_data(
-    client, stixify_file, api_schema, navigator_data
-):
-    post = models.File.objects.get(pk="dcbeb240-8dd6-4892-8e9e-7b6bda30e454")
-    post.txt2stix_data = {"navigator_layer": navigator_data}
-    post.save()
-
-    resp = client.get(
-        f"/api/v1/files/dcbeb240-8dd6-4892-8e9e-7b6bda30e454/attack-navigator/ics/",
-        data=None,
-        content_type="application/json",
-    )
-    assert resp.status_code == 200, resp.content
-    assert resp.data == {
-        "version": "4.5",
-        "name": "Basic ICS Layer",
-        "domain": "ics-attack",
-        "description": "A simple example ICS layer using only required fields.",
-        "techniques": [{"techniqueID": "T0887"}, {"techniqueID": "T0853"}],
-    }
-    api_schema["/api/v1/files/{file_id}/attack-navigator/{attack_domain}/"][
-        "GET"
-    ].validate_response(Transport.get_st_response(resp))
-
-    resp = client.get(
-        f"/api/v1/files/dcbeb240-8dd6-4892-8e9e-7b6bda30e454/attack-navigator/mobile/",
-        data=None,
-        content_type="application/json",
-    )
-    assert resp.status_code == 200, resp.content
-    assert resp.data == {
-        "version": "4.5",
-        "name": "Minimal Mobile Layer",
-        "domain": "mobile-attack",
-        "description": "A basic mobile layer with minimal configuration.",
-        "techniques": [{"techniqueID": "T1406"}, {"techniqueID": "T1475"}],
-    }
-    api_schema["/api/v1/files/{file_id}/attack-navigator/{attack_domain}/"][
-        "GET"
-    ].validate_response(Transport.get_st_response(resp))
-
-    resp = client.get(
-        f"/api/v1/files/dcbeb240-8dd6-4892-8e9e-7b6bda30e454/attack-navigator/enterprise/",
-        data=None,
-        content_type="application/json",
-    )
-    assert resp.status_code == 404, resp.content
-    assert json.loads(resp.content)["details"] == {
-        "error": "no navigator for this domain",
-        "domains": ["ics", "mobile"],
-    }
-    api_schema["/api/v1/files/{file_id}/attack-navigator/{attack_domain}/"][
-        "GET"
-    ].validate_response(Transport.get_st_response(resp))
